Improve compatibility with django 1.8
